controle de acesso e cadastro e login

+++ app/Http/Controllers/TarefasController.php
@@ -4,13 +4,14 @@ namespace App\Http\Controllers;
 
 use App\Models\Tarefa;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 
 class TarefasController extends Controller
 {
     public function index()
     {
-        $tarefas = Tarefa::latest('created_at')->get();
+        $tarefas = Tarefa::where('user_id', Auth::id())->latest()->get();
         return view('tarefas.index', compact('tarefas'));
     }
 
@@ -28,18 +29,24 @@ class TarefasController extends Controller
             'data_finalizacao' => ['nullable', 'date'],
         ]);
 
-        Tarefa::create($validated);
+        $dados = $request->all();
+        $dados['user_id'] = Auth::id();
+        Tarefa::create($dados);
 
         return redirect()->route('tarefas.index');
     }
 
     public function edit(Tarefa $tarefa)
     {
+        if ($tarefa->user_id !== Auth::id())
+            abort(403, 'Acesso negado.');
         return view('tarefas.edit', compact('tarefa'));
     }
 
     public function update(Request $request, Tarefa $tarefa)
     {
+        if ($tarefa->user_id !== Auth::id())
+            abort(403, 'Acesso negado.');
         $validated = $request->validate([
             'nome' => ['required', 'string', 'min:3', 'max:255'],
             'descricao' => ['nullable', 'string'],
@@ -53,6 +60,9 @@ class TarefasController extends Controller
 
     public function destroy(Tarefa $tarefa)
     {
+        if ($tarefa->user_id !== Auth::id())
+            abort(403, 'Acesso negado.');
+
         $tarefa->delete();
         return redirect()->route('tarefas.index');
     }
